+++ src/runtime/Utility.hpp
@@ -19,6 +19,8 @@ namespace sarus {
 namespace runtime {
 namespace utility {
 
+void setupSignalProxying(const pid_t childPid);
+
 void logMessage(const boost::format&, common::LogLevel,
                 std::ostream& out=std::cout, std::ostream& err=std::cerr);
 void logMessage(const std::string&, common::LogLevel,