+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added support for passing command-line options to `mksquashfs` through the `mksquashfsOptions` parameter in the `sarus.json` configuration file
 - Added CI unit and integration tests from source on Rocky Linux 8 and 9
+- Added explicit forwarding of standard signals from engine to OCI runtime
 
 ### Changed